srcs/backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.5
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
phonenumbers==8.13.54
psycopg==3.2.4
psycopg-binary==3.2.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pypng==0.20220715.0
pytz==2025.1
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
rpds-py==0.22.3
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
zope.interface==7.2
pillow==11.1.0
pyotp==2.3.0
pytz==2025.1
